fix: Fix export of tables and graphs [PT-175219669]
This adds a query parameter to CODAP that the CFM already looks for that causes CODAP exports to post a message with the contents and mimetype to export to the parent window.

The missing part on the SageModeler side (the parent window) was to add an event listener to cause the top level SageModeler CFM to show the export dialog.